requirements.txt
amqp==5.0.6
asgiref==3.3.3
billiard==3.6.4.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-debug-toolbar==3.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httpie==2.4.0
httplib2==0.19.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
prompt-toolkit==3.0.18
protobuf==3.15.7
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
